Move /pkg to /internal
* Moving all the current packages until `/pkg` to `/internal`
after majority of the maintainers (@skriss @youngnick @LukeShu)
prefer all packages should be internal until their APIs are deemed
stable enough to be exported and should be moved to `/pkg` then
