+++ internal/envoygateway/config/testdata/decoder/in/kube-provider.yaml
