+++ internal/envoygateway/config/testdata/decoder/in/no-api-version.yaml
