+++ internal/envoygateway/config/testdata/decoder/in/no-kind.yaml
