+++ internal/provider/kubernetes/config/crd/kustomization.yaml
