+++ internal/provider/kubernetes/config/crd/kustomizeconfig.yaml
