+++ internal/provider/kubernetes/config/crd/patches/cainjection_in_envoyproxies.yaml
