+++ internal/provider/kubernetes/config/crd/patches/webhook_in_envoyproxies.yaml
