+++ internal/provider/kubernetes/config/default/auth_proxy_patch.yaml
