+++ internal/provider/kubernetes/config/envoy-gateway/deploy_and_ns.yaml
