+++ internal/provider/kubernetes/config/prometheus/kustomization.yaml
