+++ internal/provider/kubernetes/config/rbac/auth_proxy_client_clusterrole.yaml
