+++ internal/provider/kubernetes/config/rbac/auth_proxy_role_binding.yaml
