+++ internal/provider/kubernetes/config/rbac/auth_proxy_service.yaml
