+++ internal/provider/kubernetes/config/rbac/envoyproxy_editor_role.yaml
