+++ internal/provider/kubernetes/config/rbac/envoyproxy_viewer_role.yaml
