+++ internal/provider/kubernetes/config/rbac/leader_election_role.yaml
