+++ internal/provider/kubernetes/config/samples/config_v1alpha1_envoyproxy.yaml
