+++ internal/provider/provider.go
@@ -6,8 +6,8 @@ import (
 	ctrl "sigs.k8s.io/controller-runtime"
 
 	"github.com/envoyproxy/gateway/api/config/v1alpha1"
-	"github.com/envoyproxy/gateway/pkg/envoygateway/config"
-	"github.com/envoyproxy/gateway/pkg/provider/kubernetes"
+	"github.com/envoyproxy/gateway/internal/envoygateway/config"
+	"github.com/envoyproxy/gateway/internal/provider/kubernetes"
 )
 
 func Start(svr *config.Server) error {
