+++ internal/xds/translator/context.go
