+++ internal/xds/translator/in/xds-ir/http-route.yaml
