+++ internal/xds/translator/testdata/in/xds-ir/http-route.yaml
